Move non-components to `common` folder
