--- src/data/index.js
@@ -1,5 +1,0 @@
-import TileData from './TileData';
-// export class as key and value pair in an object
-export {
-  TileData
-}
